+++ test/mailers/previews/user_mailer_preview.rb
@@ -5,4 +5,14 @@ class UserMailerPreview < ActionMailer::Preview
     UserMailer.contact("FirstName LastName", "example@test.com", "This is my message\nOn two lines!")
   end
 
+  def user_added_to_project
+    project_user = ProjectUser.where.not( invite_email: nil ).first
+    UserMailer.user_added_to_project(project_user)
+  end
+
+  def invite_user_to_project
+    project_user = ProjectUser.where.not( invite_email: nil ).first
+    UserMailer.invite_user_to_project(project_user)
+  end
+
 end